+++ Client/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,10 +1,13 @@
-## 개요
+
+## PR타이틀
 내용을 적어주세요
 
-##작업사항
+## PR내용
+
 -[]
 -[]
 -[]
-##변경로직
--내용을 적어주세요
+
+## PR생성날짜
+-YYYY/MM/DD
