+++ Client/src/index.css
@@ -1,27 +1,4 @@
-:root{
-  --main-001: #B0DBFC;
-  --main-002: #66BFFF;
-  --main-003: #62A3F4;
-  --button-001: #62A3F4;
-  --button-text: #FDFFEC;
-  --button-clicked: #486AC2;
-  --text-warn: #FF4848;
-  --text-black: #0C0D0E;
-  --text-grey:#828282;
-  --text-white:#FFFFFF;
-  --text-green:#81EE7F;
-  --border-001: #E3E6E8;
-  --border-002: #D6D9DC;
-  --border-003: #C8CCD0;
-  --font-size-sm: 14px;
-  --font-size-md: 18px;
-  --font-size-lg: 22px;
-  --font-size-h1: 34px;
-  --font-size-h2: 30px;
-  --font-size-h3: 26px;
-  }
-  
-  body {
+ body {
     margin: 0;
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
       'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
